+++ src/util/lang/bip39_korean.h
